seperating codeeditor window changes
