+++ src/components/CodeEditorWindow/index.jsx
@@ -5,7 +5,7 @@ import {
   btnconfig,
   GetCodeBlocks,
   selectBlock,
-} from "./config/Rules";
+} from "../EditorCommon/config/Rules";
 import { useClient } from "../../context/client-context";
 import { useTheme } from "@mui/material/styles";
 import { Autocomplete } from "./config/Autocomplete";
@@ -45,7 +45,6 @@ const CodeEditorWindow = ({ onChange, code, onChangeResult }) => {
 
   function handleEditorDidMount(editor, monaco) {
     editorRef.current = editor;
-    monacoRef.current = monaco;
     var decorations = [];
 
     runBtnCommandId = editor.addCommand(
