+++ src/components/CodeEditorWindow/parser.test.js
@@ -1,4 +1,4 @@
-import { GetCodeBlocks, selectBlock } from "./config/Rules";
+import { GetCodeBlocks, selectBlock } from "../EditorCommon/config/Rules";
 import { codeParse } from "./config/RequesFromCode";
 import { describe, it, expect } from "vitest";
 let testCode = `GET collections
